+++ app/src/main/java/com/calikot/mysavingquest/component/setup/recurringbills/domain/models/RecurringBillItem.kt
@@ -15,7 +15,7 @@ data class RecurringBillItem(
     val date: String,
 
     @SerialName("bill_amount")
-    val amount: Int,
+    val amount: Float,
 
     @SerialName("bill_is_auto")
     val isAuto: Boolean
